Separate auth requests from the TelegramClient
